feat(playlists, discover): implement collapsible sections for playlists and tracks with toggle functionality
