+++ mobile/lib/features/streams/presentation/screens/discover_screen.dart
@@ -23,6 +23,7 @@ import '../../../recommendations/presentation/widgets/recommended_track_tile.dar
 import '../../../tracks/domain/entities/public_track.dart';
 import '../../domain/entities/live_stream.dart';
 import '../providers/discover_notifier.dart';
+import '../../../../core/widgets/collapsible_section.dart';
 import '../../../../core/widgets/message_view.dart';
 import '../../../../core/widgets/search_field.dart';
 import '../widgets/stream_tile.dart';
@@ -309,13 +310,13 @@ class _DiscoverViewState extends State<_DiscoverView> {
         // « Pour toi » (US-09-04, déplacée dans « Découvrir » en STR-250) :
         // masquée sans contenu, placée au-dessus des pistes publiques.
         if (recos.isNotEmpty) ...[
-          _CollapsibleHeader(
+          CollapsibleHeader(
             title: 'Pour toi',
             count: recos.length,
             expanded: _recosExpanded,
             onToggle: () => setState(() => _recosExpanded = !_recosExpanded),
           ),
-          _CollapsibleContent(
+          CollapsibleContent(
             expanded: _recosExpanded,
             child: Column(
               children: [
@@ -338,13 +339,13 @@ class _DiscoverViewState extends State<_DiscoverView> {
           const SizedBox(height: 16),
         ],
         if (publicTracks.isNotEmpty) ...[
-          _CollapsibleHeader(
+          CollapsibleHeader(
             title: 'Pistes publiques',
             count: publicTracks.length,
             expanded: _publicExpanded,
             onToggle: () => setState(() => _publicExpanded = !_publicExpanded),
           ),
-          _CollapsibleContent(
+          CollapsibleContent(
             expanded: _publicExpanded,
             child: Column(
               children: [
@@ -470,99 +471,6 @@ class _CategoryChip extends StatelessWidget {
             ),
           ),
         ),
-      ),
-    );
-  }
-}
-
-/// En-tête de section repliable : titre + compteur discret + chevron qui pivote.
-/// Toute la ligne est tappable.
-class _CollapsibleHeader extends StatelessWidget {
-  const _CollapsibleHeader({
-    required this.title,
-    required this.count,
-    required this.expanded,
-    required this.onToggle,
-  });
-
-  final String title;
-  final int count;
-  final bool expanded;
-  final VoidCallback onToggle;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = Theme.of(context).colorScheme;
-    final text = Theme.of(context).textTheme;
-
-    return Semantics(
-      button: true,
-      label: '$title, ${expanded ? 'réduire' : 'développer'} la section',
-      onTap: onToggle,
-      excludeSemantics: true,
-      child: InkWell(
-        onTap: onToggle,
-        borderRadius: BorderRadius.circular(12),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(vertical: 6),
-          child: Row(
-            children: [
-              Text(
-                title,
-                style: text.titleMedium?.copyWith(fontWeight: FontWeight.w700),
-              ),
-              const SizedBox(width: 8),
-              Container(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 8, vertical: 2),
-                decoration: BoxDecoration(
-                  color: colors.surfaceContainerHighest,
-                  borderRadius: BorderRadius.circular(10),
-                ),
-                child: Text(
-                  '$count',
-                  style: text.labelSmall?.copyWith(
-                    color: colors.onSurfaceVariant,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-              ),
-              const Spacer(),
-              AnimatedRotation(
-                turns: expanded ? 0.5 : 0.0,
-                duration: const Duration(milliseconds: 220),
-                curve: Curves.easeInOut,
-                child: Icon(
-                  Icons.keyboard_arrow_down,
-                  color: colors.onSurfaceVariant,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Contenu d'une section repliable : anime la hauteur entre le contenu complet
-/// et zéro, avec un fondu, quand [expanded] change.
-class _CollapsibleContent extends StatelessWidget {
-  const _CollapsibleContent({required this.expanded, required this.child});
-
-  final bool expanded;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedSize(
-      duration: const Duration(milliseconds: 220),
-      curve: Curves.easeInOut,
-      alignment: Alignment.topCenter,
-      child: AnimatedOpacity(
-        opacity: expanded ? 1 : 0,
-        duration: const Duration(milliseconds: 180),
-        child: expanded ? child : const SizedBox(width: double.infinity),
       ),
     );
   }
